reorg test fixtures
