+++ Python/qdk-chemistry/qdk_chemistry/tests/test_nwchem.py
@@ -2,25 +2,8 @@ import pytest
 from unittest import mock
 
 from qdk_chemistry.solvers.nwchem import create_input_deck, format_geometry_from_mol
-from qdk_chemistry.geometry import Geometry, Element
 
 from rdkit.Chem import AllChem as Chem
-
-
-@pytest.fixture()
-def h2o():
-    mol = Chem.AddHs(Chem.MolFromSmiles("O"))
-    return mol
-
-
-@pytest.fixture()
-def geometry(h2o):
-    el = [
-        Element("O", 0.002, 0.398, 0.0),
-        Element("H", 0.762, -0.203, 0.0),
-        Element("H", -0.764, -0.195, 0.0)
-    ]
-    return Geometry(el, charge=Chem.GetFormalCharge(h2o))
 
 
 @pytest.fixture()
